\name{bootruin-package}
\alias{bootruin-package}
\docType{package}
\title{A Bootstrap Test for the Probability of Ruin in the Compound Poisson Risk process}

\description{
This package provides a bootstrap test for the probability of ruin in the classical (compound
Poisson) risk process and some procedures for comparing the performance of the bootstrap test and
the test using the asymptotic normal approximation.
}

\details{
See the reference for more information.
}

\references{
Baumgartner, B. and Gatto, R. (2010) \emph{A Bootstrap Test for the Probability of Ruin in the
Compound Poisson Risk Process}. ASTIN Bulletin, \bold{40}(1), pp. 241--255.
}


\author{
Benjamin Baumgartner \email{benjamin.baumgartner@stat.unibe.ch},\cr
Riccardo Gatto \email{gatto@stat.unibe.ch}
}

\keyword{package}
